refactor: move os dependent code to another code
